Updated from trunk changes. (cherry picked from commit 4567aa3280933e6f4584969b4eed6964d8191f75)
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@180473 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/j3/VMCore/JavaClass.h
@@ -84,7 +84,7 @@ public:
   // createElementValue - create the Java type associated with the value
   // of the current annotation key.
   //
-  JavaObject* createElementValue(bool nextParameterIsTypeOfMethod,JavaObject* type, const UTF8* lastKey);
+  JavaObject* createElementValue(bool nextParameterIsTypeOfMethod, JavaObject* type, const UTF8* lastKey);
 
   void fillArray(JavaObject* res, int numValues, UserClassArray* classArray);
 };
@@ -464,7 +464,7 @@ public:
 
   /// nbVirtualMethods - The number of virtual methods.
   ///
-  uint32 nbVirtualMethods;
+  uint16 nbVirtualMethods;
   
   /// staticMethods - List of all the static methods defined by this class.
   ///
